1. 在package.json中增加升级bone的指令 2. 优化了文件是否存在的判断

+++ generators/templates/.bone/src/site/middleware/weaver/jsonFile.js
@@ -4,9 +4,7 @@ var initFile = function(filepath, initValue) {
     throw new Error('filepath is unvalued');
     return;
   }
-  try {
-    fs.readFileSync(filepath);
-  } catch (e) {
+  if (!fs.existsSync(filepath)) {
     fs.openSync(filepath, 'w+');
     if (initValue) {
       fs.writeFileSync(filepath, JSON.stringify(initValue, null, '   '), 'UTF-8');
@@ -14,7 +12,7 @@ var initFile = function(filepath, initValue) {
   }
 };
 var readJsonFile = function(filepath) {
-  try { 
+  try {
     var r = fs.readFileSync(filepath);
   } catch (e) {
     console.error(e);
